Add a regression spec to cover rails case
In 2.2.10 we introduced an issue where Rails Gemfile could no longer be
bundled.

Cover that case to make sure we don't regress there.

Note that current `disable_multisource` logic breaks this, so this will
require further work, but at least we make sure that the changes to
enable separate GEM lockfile sections don't break this case.